+++ src/stapi_fastapi/models/root.py
@@ -1,7 +1,11 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 from stapi_fastapi.models.shared import Link
 
 
 class RootResponse(BaseModel):
-    links: list[Link]
+    id: str
+    conformsTo: list[str] = Field(default_factory=list)
+    title: str = ""
+    description: str = ""
+    links: list[Link] = Field(default_factory=list)